backporting some fixes from the ros1 version of the repo

+++ sdk/src/arch/linux/net_socket.cpp
@@ -19,7 +19,7 @@
 namespace rp{ namespace net {
 
 
-static inline int _halAddrTypeToOSType(SocketAddress::address_type_t type) 
+static inline int _halAddrTypeToOSType(SocketAddress::address_type_t type)
 {
     switch (type) {
         case SocketAddress::ADDRESS_TYPE_INET:
@@ -36,11 +36,11 @@ static inline int _halAddrTypeToOSType(SocketAddress::address_type_t type)
 }
 
 
-SocketAddress::SocketAddress() 
+SocketAddress::SocketAddress()
 {
     _platform_data = reinterpret_cast<void *>(new sockaddr_storage);
     memset(_platform_data, 0, sizeof(sockaddr_storage));
-    
+
     reinterpret_cast<sockaddr_storage *>(_platform_data)->ss_family = AF_INET;
 }
 
@@ -56,7 +56,7 @@ SocketAddress::SocketAddress(const char * addrString, int port, SocketAddress::a
 {
     _platform_data = reinterpret_cast<void *>(new sockaddr_storage);
     memset(_platform_data, 0, sizeof(sockaddr_storage));
-    
+
     // default to ipv4 in case the following operation fails
     reinterpret_cast<sockaddr_storage *>(_platform_data)->ss_family = AF_INET;
 
@@ -127,8 +127,8 @@ u_result SocketAddress::setAddressFromString(const char * address_string,  Socke
     switch (type) {
         case ADDRESS_TYPE_INET:
             reinterpret_cast<sockaddr_storage *>(_platform_data)->ss_family = AF_INET;
-            ans = inet_pton(AF_INET, 
-                            address_string, 
+            ans = inet_pton(AF_INET,
+                            address_string,
                             &reinterpret_cast<sockaddr_in *>(_platform_data)->sin_addr);
         break;
 
@@ -136,8 +136,8 @@ u_result SocketAddress::setAddressFromString(const char * address_string,  Socke
         case ADDRESS_TYPE_INET6:
             
             reinterpret_cast<sockaddr_storage *>(_platform_data)->ss_family = AF_INET6;
-            ans = inet_pton(AF_INET6, 
-                            address_string, 
+            ans = inet_pton(AF_INET6,
+                            address_string,
                             &reinterpret_cast<sockaddr_in6  *>(_platform_data)->sin6_addr);
         break;
 
@@ -167,7 +167,7 @@ u_result SocketAddress::getAddressAsString(char * buffer, size_t buffersize) con
 
         break;
     }
-    return ans<=0?RESULT_OPERATION_FAIL:RESULT_OK;
+    return ((uintptr_t)ans)<=0?RESULT_OPERATION_FAIL:RESULT_OK;
 }
 
 
@@ -184,7 +184,6 @@ size_t SocketAddress::LoopUpHostName(const char * hostname, const char * sevicen
 
     if (!performDNS) {
         hints.ai_family |= AI_NUMERICSERV | AI_NUMERICHOST;
-    
     }
 
     ans = getaddrinfo(hostname, sevicename, &hints, &result);
@@ -196,7 +195,6 @@ size_t SocketAddress::LoopUpHostName(const char * hostname, const char * sevicen
         return 0;
     }
 
-    
     for (struct addrinfo * cursor = result; cursor != NULL; cursor = cursor->ai_next) {
         if (cursor->ai_family == ADDRESS_TYPE_INET || cursor->ai_family == ADDRESS_TYPE_INET6) {
             sockaddr_storage * storagebuffer = new sockaddr_storage;
@@ -206,7 +204,6 @@ size_t SocketAddress::LoopUpHostName(const char * hostname, const char * sevicen
         }
     }
 
-    
     freeaddrinfo(result);
 
     return addresspool.size();
@@ -221,7 +218,6 @@ u_result SocketAddress::getRawAddress(_u8 * buffer, size_t bufferSize) const
 
             memcpy(buffer, &reinterpret_cast<const sockaddr_in *>(_platform_data)->sin_addr.s_addr, sizeof(reinterpret_cast<const sockaddr_in *>(_platform_data)->sin_addr.s_addr));
 
-            
             break;
         case ADDRESS_TYPE_INET6:
             if (bufferSize < sizeof(in6_addr::s6_addr)) return RESULT_INSUFFICIENT_MEMORY;
@@ -309,7 +305,7 @@ void SocketAddress::setAnyAddress(SocketAddress::address_type_t type)
 ///--------------------------------
 
 
-namespace rp { namespace arch { namespace net{ 
+namespace rp { namespace arch { namespace net{
 
 using namespace rp::net;
 
@@ -328,7 +324,7 @@ public:
         this->setTimeout(DEFAULT_SOCKET_TIMEOUT, SOCKET_DIR_BOTH);
     }
 
-    virtual ~StreamSocketImpl() 
+    virtual ~StreamSocketImpl()
     {
         close(_socket_fd);
     }
@@ -369,8 +365,8 @@ public:
     {
         int ans;
         timeval tv;
-        tv.tv_sec = timeout / 1000; 
-        tv.tv_usec = (timeout % 1000) * 1000; 
+        tv.tv_sec = timeout / 1000;
+        tv.tv_usec = (timeout % 1000) * 1000;
 
         if (msk & SOCKET_DIR_RD) {
              ans = ::setsockopt( _socket_fd, SOL_SOCKET, SO_RCVTIMEO, &tv, sizeof(tv) );
@@ -413,7 +409,7 @@ public:
         return ans?RESULT_OPERATION_FAIL:RESULT_OK;
     }
 
-    virtual StreamSocket * accept(SocketAddress * pairAddress) 
+    virtual StreamSocket * accept(SocketAddress * pairAddress)
     {
         size_t addrsize;
         addrsize = sizeof(sockaddr_storage);
@@ -432,7 +428,7 @@ public:
         return waitforData(timeout);
     }
 
-    virtual u_result send(const void * buffer, size_t len) 
+    virtual u_result send(const void * buffer, size_t len)
     {
         size_t ans = ::send( _socket_fd, buffer, len, MSG_NOSIGNAL);
         if (ans == (int)len) {
@@ -467,8 +463,6 @@ public:
                 default:
                     return RESULT_OPERATION_FAIL;
             }
-
-            
 
         } else {
             recv_len = ans;
